refactor: extract shared HRT-duration and next-dose helpers
HrtDurationL10n.localizedText replaces the same duration -> localized
string switch that was duplicated in HrtWidgetService and
IntakeCounterCard.

NextDoseResolver pulls the "which schedule is next due" urgency
ranking and timing-label logic out of NextDoseWidgetService into a
pure, reusable class. Both are prep for the upcoming Status widget,
which needs the exact same duration text and next-dose picking the
other widgets already compute -- extracting first means it can reuse
them instead of drifting out of sync with a third copy.
